Feature/refactor background controller (#425)
* refactor: add function to display popUp

* fix: some small stuffs

* Added methods to connection module

* Added transaction module

* Exported new module methods

* Added InformationModule

* Added information module to controller

* Minor merge error fixes

* Addd sign methods

* Lint fixes

* Lint fixes

* Exported missing methods

* Fixed static method error
