+++ source/Modules/Controller/index.js
@@ -1,1 +1,3 @@
 export * from './connection';
+export * from './transaction';
+export * from './information';
