+++ source/Modules/index.js
@@ -1,2 +1,4 @@
 export * from './storageManager';
 export * from './Controller/connection';
+export * from './Controller/transaction';
+export * from './Controller/information';
